reduced struct size by using different depth

+++ src/config.rs
@@ -10,7 +10,7 @@ pub struct SearchConfig {
     pub keep_dirs: bool,
     pub extension_match: Option<Arc<[u8]>>,
     pub file_name: bool,
-    pub depth: Option<u16>,
+    pub depth: Option<u8>,
 }
 
 impl SearchConfig {
@@ -23,7 +23,7 @@ impl SearchConfig {
         keep_dirs: bool,
         file_name: bool,
         extension_match: Option<Arc<[u8]>>,
-        depth: Option<u16>,
+        depth: Option<u8>,
     ) -> Self {
         let reg = if pattern == "." || pattern.is_empty() {
             None
